bot/requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
appdirs==1.4.4
astroid==2.4.1
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
config==0.5.0.post0
discord.py==1.7.3
distlib==0.3.1
dodgy==0.2.1
filelock==3.0.12
flake8==3.8.4
flake8-polyfill==1.0.2
idna==2.10
isort==4.3.21
kombu==5.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.23.1
multidict==5.1.0
numpy==1.19.5
pep8-naming==0.10.0
pipenv==2020.11.15
prompt-toolkit==3.0.20
prospector==1.3.1
pycodestyle==2.6.0
pydocstyle==6.1.1
pyflakes==2.2.0
pygame==2.0.1
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pymongo==3.12.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requirements-detector==0.7
setoptconf==0.2.0
six==1.16.0
snowballstemmer==2.1.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.2
vine==5.0.0
virtualenv==20.4.6
virtualenv-clone==0.5.4
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
